requirements.txt
gitdb==0.6.4
GitPython==0.3.6
requests==2.5.1
smmap==0.9.0
